Move Config related tests from keymap tests Config tests
